게임로비 진입시 UDP소켓 생성 서버로 포트 전달 서버에서 클라의 엔드포인트 저장 플레이어 생성 후 저장, 플레이어 아이디 전달

+++ GameServerContent/GameSession.cpp
@@ -82,11 +82,11 @@ void GameSession::BroadCastMachingPacket()
 {
 	// ContentManager에서 락을 걸어주므로 걸 필요없다
 	GameMatchPacket packet;
-	packet.sessionIdx = m_sessionIdx;
+	packet.sessionIdx = static_cast<int>(m_sessionIdx);
 
 	for (size_t i = 0; i < m_userInfo.size(); i++)
 	{
-		packet.userIdx = i;
+		packet.userIdx = static_cast<int>(i);
 
 		ServerSerializer sr;
 		packet >> sr;
